backend/requirements.txt
# ============================================
# NexusAI Backend - Enterprise Dependencies
# ============================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
jwt==1.3.1
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Async & Concurrency
aiofiles==23.2.1
httpx==0.25.1

# Code Execution & Analysis
GitPython==3.1.40
bandit==1.7.5
pylint==3.0.3
black==23.12.1
requests==2.31.0

# AI Integration
litellm==1.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
coverage==7.3.2

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Utilities
pydantic-extra-types==2.0.0
slowapi==0.1.9  # Rate limiting
python-dateutil==2.8.2

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
